requirements.txt
alabaster==0.7.6
appnope==0.1.0
Babel==2.1.1
backports-abc==0.4
backports.ssl-match-hostname==3.4.0.2
certifi==2015.11.20.1
cycler==0.9.0
decorator==4.0.4
dj-database-url==0.3.0
Django==1.9.1
django-braces==1.8.1
django-dotenv==1.3.0
django-gravatar2==1.4.0
docutils==0.12
et-xmlfile==1.0.1
functools32==3.2.3.post2
gnureadline==6.3.3
ipykernel==4.2.0
ipyparallel==4.1.0
ipython==4.0.1
ipython-genutils==0.1.0
jdcal==1.2
Jinja2==2.8
jsonschema==2.5.1
jupyter-client==4.1.1
jupyter-core==4.0.6
MarkupSafe==0.23
matplotlib==1.5.0
mistune==0.7.1
nbconvert==4.1.0
nbformat==4.0.1
nose==1.3.7
notebook==4.0.6
numpy==1.10.1
numpydoc==0.5
openpyxl==2.3.1
pandas==0.17.1
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.5
psycopg2==2.6.1
ptyprocess==0.5
Pygments==2.0.2
pyparsing==2.0.6
python-dateutil==2.4.2
pytz==2015.7
pyzmq==15.1.0
qtconsole==4.1.1
requests==2.8.1
scipy==0.16.1
seaborn==0.6.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.0
Sphinx==1.3.3
sphinx-rtd-theme==0.1.9
terminado==0.5
testpath==0.2
tornado==4.3
traitlets==4.0.0
wheel==0.24.0
whitenoise==2.0.6
